requirements.txt
dateutils==0.6.6
Django==2.1
djangorestframework==3.8.2
python-dateutil==2.7.3
pytz==2018.5
six==1.11.0
